examples/gcf-pubsub-vm-delete-event-handler/requirements.txt
google-api-python-client>=1.7.9
python-json-logger>=0.1.11
google-cloud-logging>=1.11.0
